server/requirements.txt
Django==1.8
gunicorn==19.1.1
python-dateutil==2.5.1
celery==3.1.23
django-uuidfield==0.5.0 # required by push-notifications
django-push-notifications==1.4.1
